refactor: extract Messages to a separate component for better readability
